rename runtimes to explicit vlm_runtimes
