+++ docling/models/runtimes/api_openai_compatible_vlm_runtime.py
