+++ docling/models/runtimes/transformers_vlm_runtime.py
