+++ docling/models/runtimes/vllm_vlm_runtime.py
